- Removed partial configuration - Revamped application-wide navigation
