--- .kahi-docs/partials.config.ts
@@ -1,9 +1,0 @@
-import {define_partials} from "@kahi-docs/config";
-
-import Footer from "./partials/Footer.svelte";
-import Landing from "./partials/Landing.svelte";
-
-export default define_partials({
-    Footer,
-    Landing,
-});
