+++ packages/@kahi-docs-config/src/index.ts
@@ -3,5 +3,4 @@ export * from "./types/navigation";
 
 export * from "./config/application";
 export * from "./config/navigation";
-export * from "./config/partials";
 export * from "./config/repl";
