+++ packages/@kahi-docs-shared/package.json
@@ -20,6 +20,7 @@
         "test:types": "tsc --noEmit"
     },
     "devDependencies": {
+        "@sveltejs/kit": "^1.0.0-next.190",
         "prettier": "^2.3.1",
         "typescript": "^4.3.5"
     },
