+++ packages/@kahi-docs-shared/src/index.ts
@@ -1,12 +1,12 @@
 export * from "./stores/applicationconfig";
 export * from "./stores/docsrender";
 export * from "./stores/navigation";
-export * from "./stores/partialsconfig";
 export * from "./stores/preferences";
 export * from "./stores/snippet";
 
 export * from "./util/functional";
 export * from "./util/html";
+export * from "./util/page";
 export * from "./util/string";
 export * from "./util/svelte";
 export * from "./util/url";
