+++ packages/@kahi-docs-web/src/routes/api/v2/configuration/application.json.ts
@@ -2,7 +2,7 @@ import type {RequestHandler} from "@sveltejs/kit";
 
 import type {IApplicationConfigGet} from "../../../../shared/api";
 
-import APPLICATION_CONFIG from "../../../../../.kahi-docs/application.config";
+import APPLICATION_CONFIG from "../../../../../../../.kahi-docs/application.config";
 
 export const get: RequestHandler = (request) => {
     return {
